+++ uniffi_macros/src/object.rs
@@ -3,7 +3,9 @@ use quote::quote;
 use syn::DeriveInput;
 use uniffi_meta::free_fn_symbol_name;
 
-use crate::util::{create_metadata_items, ident_to_string, mod_path, tagged_impl_header};
+use crate::util::{
+    create_metadata_items, derive_ffi_traits, ident_to_string, mod_path, tagged_impl_header,
+};
 
 pub fn expand_object(input: DeriveInput, udl_mode: bool) -> syn::Result<TokenStream> {
     let module_path = mod_path()?;
@@ -20,15 +22,11 @@ pub fn expand_object(input: DeriveInput, udl_mode: bool) -> syn::Result<TokenStr
         #[doc(hidden)]
         #[no_mangle]
         pub extern "C" fn #free_fn_ident(
-            ptr: *const ::std::ffi::c_void,
+            handle: ::uniffi::Handle,
             call_status: &mut ::uniffi::RustCallStatus
         ) {
             uniffi::rust_call(call_status, || {
-                assert!(!ptr.is_null());
-                let ptr = ptr.cast::<#ident>();
-                unsafe {
-                    ::std::sync::Arc::decrement_strong_count(ptr);
-                }
+                <#ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::remove(handle);
                 Ok(())
             });
         }
@@ -42,6 +40,7 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
     let name = ident_to_string(ident);
     let impl_spec = tagged_impl_header("FfiConverterArc", ident, udl_mode);
     let lift_ref_impl_spec = tagged_impl_header("LiftRef", ident, udl_mode);
+    let derive_ffi_traits = derive_ffi_traits(ident, udl_mode, &["SlabAlloc"]);
     let mod_path = match mod_path() {
         Ok(p) => p,
         Err(e) => return e.into_compile_error(),
@@ -54,6 +53,8 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
         // and thus help the user debug why the requirement isn't being met.
         uniffi::deps::static_assertions::assert_impl_all!(#ident: Sync, Send);
 
+        #derive_ffi_traits
+
         #[doc(hidden)]
         #[automatically_derived]
         /// Support for passing reference-counted shared objects via the FFI.
@@ -62,8 +63,7 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
         /// by reference must be encapsulated in an `Arc`, and must be safe to share
         /// across threads.
         unsafe #impl_spec {
-            // Don't use a pointer to <T> as that requires a `pub <T>`
-            type FfiType = *const ::std::os::raw::c_void;
+            type FfiType = ::uniffi::Handle;
 
             /// When lowering, we have an owned `Arc` and we transfer that ownership
             /// to the foreign-language code, "leaking" it out of Rust's ownership system
@@ -75,7 +75,7 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
             /// call the destructor function specific to the type `T`. Calling the destructor
             /// function for other types may lead to undefined behaviour.
             fn lower(obj: ::std::sync::Arc<Self>) -> Self::FfiType {
-                ::std::sync::Arc::into_raw(obj) as Self::FfiType
+                <#ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::insert(obj)
             }
 
             /// When lifting, we receive a "borrow" of the `Arc` that is owned by
@@ -84,11 +84,7 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
             /// Safety: the provided value must be a pointer previously obtained by calling
             /// the `lower()` or `write()` method of this impl.
             fn try_lift(v: Self::FfiType) -> ::uniffi::Result<::std::sync::Arc<Self>> {
-                let v = v as *const #ident;
-                // We musn't drop the `Arc` that is owned by the foreign-language code.
-                let foreign_arc = ::std::mem::ManuallyDrop::new(unsafe { ::std::sync::Arc::<Self>::from_raw(v) });
-                // Take a clone for our own use.
-                Ok(::std::sync::Arc::clone(&*foreign_arc))
+                Ok(<#ident as ::uniffi::SlabAlloc<crate::UniFfiTag>>::get_clone(v))
             }
 
             /// When writing as a field of a complex structure, make a clone and transfer ownership
@@ -100,8 +96,7 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
             /// call the destructor function specific to the type `T`. Calling the destructor
             /// function for other types may lead to undefined behaviour.
             fn write(obj: ::std::sync::Arc<Self>, buf: &mut Vec<u8>) {
-                ::uniffi::deps::static_assertions::const_assert!(::std::mem::size_of::<*const ::std::ffi::c_void>() <= 8);
-                ::uniffi::deps::bytes::BufMut::put_u64(buf, <Self as ::uniffi::FfiConverterArc<crate::UniFfiTag>>::lower(obj) as u64);
+                ::uniffi::deps::bytes::BufMut::put_i64(buf, <Self as ::uniffi::FfiConverterArc<crate::UniFfiTag>>::lower(obj).as_raw())
             }
 
             /// When reading as a field of a complex structure, we receive a "borrow" of the `Arc`
@@ -110,9 +105,10 @@ pub(crate) fn interface_impl(ident: &Ident, udl_mode: bool) -> TokenStream {
             /// Safety: the buffer must contain a pointer previously obtained by calling
             /// the `lower()` or `write()` method of this impl.
             fn try_read(buf: &mut &[u8]) -> ::uniffi::Result<::std::sync::Arc<Self>> {
-                ::uniffi::deps::static_assertions::const_assert!(::std::mem::size_of::<*const ::std::ffi::c_void>() <= 8);
                 ::uniffi::check_remaining(buf, 8)?;
-                <Self as ::uniffi::FfiConverterArc<crate::UniFfiTag>>::try_lift(::uniffi::deps::bytes::Buf::get_u64(buf) as Self::FfiType)
+                <Self as ::uniffi::FfiConverterArc<crate::UniFfiTag>>::try_lift(
+                    ::uniffi::Handle::from_raw(::uniffi::deps::bytes::Buf::get_i64(buf))
+                )
             }
 
             const TYPE_ID_META: ::uniffi::MetadataBuffer = ::uniffi::MetadataBuffer::from_code(::uniffi::metadata::codes::TYPE_INTERFACE)